+++ packages/dedupe-mixin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.4.0
+
+### Minor Changes
+
+- 935c8ffe: Drop support for Node@14
+
 ## 1.3.1
 
 ### Patch Changes