+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 12.0.0
+
+### Patch Changes
+
+- Updated dependencies [14e4270a]
+  - eslint-plugin-lit-a11y@4.0.0
+
+## 11.0.0
+
+### Major Changes
+
+- 935c8ffe: Drop support for Node@14
+
+### Patch Changes
+
+- Updated dependencies [935c8ffe]
+  - eslint-plugin-lit-a11y@3.0.0
+
 ## 10.0.0
 
 ### Patch Changes