+++ packages/scoped-elements/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 2.2.0
+
+### Minor Changes
+
+- 935c8ffe: Drop support for Node@14
+
+### Patch Changes
+
+- 80c6ae66: Use latest @web/\* dependencies.
+  Upgrade Rollup to `v3.x`
+- Updated dependencies [935c8ffe]
+  - @open-wc/dedupe-mixin@1.4.0
+
 ## 2.1.5
 
 ### Patch Changes